gameChess/requirements.txt
pygame==2.5.2
python-chess==1.999.0 